# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Report Analysis API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LabValueStatus =
  (typeof LabValueStatus)[keyof typeof LabValueStatus];

export const LabValueStatus = {
  normal: "normal",
  high: "high",
  low: "low",
  critical: "critical",
} as const;

export interface LabValue {
  name: string;
  value: string;
  unit: string;
  referenceRange: string;
  status: LabValueStatus;
  explanation: string;
}

export type ReportStatus = (typeof ReportStatus)[keyof typeof ReportStatus];

export const ReportStatus = {
  pending: "pending",
  analyzing: "analyzing",
  done: "done",
  error: "error",
} as const;

export interface Report {
  id: number;
  fileName: string;
  fileType: string;
  uploadedAt: string;
  status: ReportStatus;
  /** @nullable */
  reportType?: string | null;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  reportDate?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  simplifiedExplanation?: string | null;
  /** @nullable */
  healthInsights?: string | null;
  /** @nullable */
  labValues?: LabValue[] | null;
  /** @nullable */
  abnormalCount?: number | null;
  /** @nullable */
  criticalCount?: number | null;
}

export interface ReportsSummary {
  totalReports: number;
  reportsWithAbnormal: number;
  reportsWithCritical: number;
  recentReports: Report[];
}