# lib/api-zod/src/generated/types/labValue.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Report Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { LabValueStatus } from "./labValueStatus";

export interface LabValue {
  name: string;
  value: string;
  unit: string;
  referenceRange: string;
  status: LabValueStatus;
  explanation: string;
}