# lib/api-zod/src/generated/types/report.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Report Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { LabValue } from "./labValue";
import type { ReportStatus } from "./reportStatus";

export interface Report {
  id: number;
  fileName: string;
  fileType: string;
  uploadedAt: string;
  status: ReportStatus;
  /** @nullable */
  reportType?: string | null;
  /** @nullable */
  patientName?: string | null;
  /** @nullable */
  reportDate?: string | null;
  /** @nullable */
  summary?: string | null;
  /** @nullable */
  simplifiedExplanation?: string | null;
  /** @nullable */
  healthInsights?: string | null;
  /** @nullable */
  labValues?: LabValue[] | null;
  /** @nullable */
  abnormalCount?: number | null;
  /** @nullable */
  criticalCount?: number | null;
}