# lib/api-zod/src/generated/types/reportsSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Medical Report Analysis API
 * OpenAPI spec version: 0.1.0
 */
import type { Report } from "./report";

export interface ReportsSummary {
  totalReports: number;
  reportsWithAbnormal: number;
  reportsWithCritical: number;
  recentReports: Report[];
}